Get rid of workers
