+++ CMakeLists.txt
@@ -19,8 +19,6 @@ set(LIB_SOURCES
         graphsearch/query.cpp
         graphsearch/querypath.cpp
         graphsearch/blast/blastsearch.cpp
-        graphsearch/blast/buildblastdatabaseworker.cpp
-        graphsearch/blast/runblastsearchworker.cpp
         graphsearch/minimap2/minimap2search.cpp
         command_line/commoncommandlinefunctions.cpp
         command_line/image.cpp
