--- graphsearch/blast/buildblastdatabaseworker.h
@@ -1,55 +1,0 @@
-//Copyright 2017 Ryan Wick
-
-//This file is part of Bandage
-
-//Bandage is free software: you can redistribute it and/or modify
-//it under the terms of the GNU General Public License as published by
-//the Free Software Foundation, either version 3 of the License, or
-//(at your option) any later version.
-
-//Bandage is distributed in the hope that it will be useful,
-//but WITHOUT ANY WARRANTY; without even the implied warranty of
-//MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-//GNU General Public License for more details.
-
-//You should have received a copy of the GNU General Public License
-//along with Bandage.  If not, see <http://www.gnu.org/licenses/>.
-
-
-#ifndef BUILDBLASTDATABASEWORKER_H
-#define BUILDBLASTDATABASEWORKER_H
-
-#include <QTemporaryDir>
-
-class AssemblyGraph;
-class QProcess;
-
-//This class carries out the task of running makeblastdb on
-//the graph's nodes.
-//It is a separate class because when run from the GUI, this
-//process takes place in a separate thread.
-
-class BuildBlastDatabaseWorker : public QObject {
-    Q_OBJECT
-public:
-    BuildBlastDatabaseWorker(QString makeblastdbCommand, const AssemblyGraph &graph,
-                             const QTemporaryDir &workdir);
-    QString m_error;
-
-private:
-    QString m_makeblastdbCommand;
-    const AssemblyGraph &m_graph;
-    const QTemporaryDir &m_workdir;
-
-    QProcess *m_makeblastdb = nullptr;
-    bool m_cancelBuildBlastDatabase = false;
-
-public slots:
-    bool buildBlastDatabase();
-    void cancel();
-
-signals:
-    void finishedBuild(QString error);
-};
-
-#endif // BUILDBLASTDATABASEWORKER_H
